Move `Path` into its own sub-module
